TypingUI working with MT server.
Completed 'active state' implementation.
Rewrote TypingState completely.
Redesigned input textarea to contain userText instead of mtText.
Updated TypingModel and TypingUI accordingly.
Removed dependency on 'elphaba' library.
Added support for continuous keypress.
Added support for rendering underline only for the 'term' portion of a token and not for the 'separator'.
Updated colors to use Maureen Stone's palette.
Added support for a ghost caret when the user move the caret beyond the end of the current user-entered text.